==> scripts/00-install_packages.R <==
#### Preamble ####
# Purpose: Installs necessary packages to execute scripts and Paper.qmd.
# Author: Ricky Fung
# Date: 14 April 2023
# Contact: ricky.fung@mail.utoronto.ca


#### Workspace setup ####

install.packages("tidyverse")
install.packages("dplyr")
install.packages("arrow")
install.packages("janitor")
install.packages("knitr")
install.packages("rstanarm")
install.packages("modelsummary")
install.packages("png")
install.packages("ggplot2")
install.packages("reshape2")
